Создан компонент кнопки лайка LikeButton. Логика обработки лайка перенесена из хука useControlTable в компонент LikeButton.
